requirements.txt
requests
uuid
websocket
